+++ templates/components/vectordbs/typescript/chroma/generate.ts
@@ -1,7 +1,7 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 import * as dotenv from "dotenv";
 import { VectorStoreIndex, storageContextFromDefaults } from "llamaindex";
-import { AstraDBVectorStore } from "llamaindex/storage/vectorStore/AstraDBVectorStore";
+import { ChromaVectorStore } from "llamaindex/storage/vectorStore/ChromaVectorStore";
 import { getDocuments } from "./loader";
 import { initSettings } from "./settings";
 import { checkRequiredEnvVars } from "./shared";
@@ -12,23 +12,20 @@ async function loadAndIndex() {
   // load objects from storage and convert them into LlamaIndex Document objects
   const documents = await getDocuments();
 
-  // create vector store and a collection
-  const collectionName = process.env.ASTRA_DB_COLLECTION!;
-  const vectorStore = new AstraDBVectorStore();
-  await vectorStore.create(collectionName, {
-    vector: {
-      dimension: parseInt(process.env.EMBEDDING_DIM!),
-      metric: "cosine",
-    },
-  });
-  await vectorStore.connect(collectionName);
+  // create vector store
+  const chromaUri = `http://${process.env.CHROMA_HOST}:${process.env.CHROMA_PORT}`;
 
-  // create index from documents and store them in Astra
+  const vectorStore = new ChromaVectorStore({
+    collectionName: process.env.CHROMA_COLLECTION,
+    chromaClientParams: { path: chromaUri },
+  });
+
+  // create index from all the Documentss and store them in Pinecone
   console.log("Start creating embeddings...");
   const storageContext = await storageContextFromDefaults({ vectorStore });
   await VectorStoreIndex.fromDocuments(documents, { storageContext });
   console.log(
-    "Successfully created embeddings and save to your Astra database.",
+    "Successfully created embeddings and save to your ChromaDB index.",
   );
 }